Move: Superblock class under `org.sysethereum.agents.core.bridge`
