+++ src/main/java/org/sysethereum/agents/constants/EthGanacheSyscoinMainAgentConstants.java
@@ -3,7 +3,7 @@ package org.sysethereum.agents.constants;
 import com.google.common.collect.Lists;
 import org.bitcoinj.core.Sha256Hash;
 import org.sysethereum.agents.core.syscoin.Keccak256Hash;
-import org.sysethereum.agents.core.syscoin.Superblock;
+import org.sysethereum.agents.core.bridge.Superblock;
 import org.libdohj.params.SyscoinMainNetParams;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
