+++ src/main/java/org/sysethereum/agents/constants/IntegrationAgentConstants.java
@@ -4,7 +4,7 @@ import com.google.common.collect.Lists;
 import org.bitcoinj.core.Sha256Hash;
 import org.libdohj.params.SyscoinTestNet3Params;
 import org.sysethereum.agents.core.syscoin.Keccak256Hash;
-import org.sysethereum.agents.core.syscoin.Superblock;
+import org.sysethereum.agents.core.bridge.Superblock;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
