+++ src/main/java/org/sysethereum/agents/core/syscoin/SuperblockChain.java
@@ -12,6 +12,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
+import org.sysethereum.agents.core.bridge.Superblock;
 
 import javax.annotation.PostConstruct;
 import javax.annotation.PreDestroy;
