+++ src/main/java/org/sysethereum/agents/tool/GenesisSuperblockGeneratorMain.java
@@ -9,7 +9,7 @@ import org.sysethereum.agents.constants.AgentConstants;
 import org.sysethereum.agents.constants.SystemProperties;
 import org.sysethereum.agents.core.syscoin.SyscoinWrapper;
 import org.sysethereum.agents.core.syscoin.Keccak256Hash;
-import org.sysethereum.agents.core.syscoin.Superblock;
+import org.sysethereum.agents.core.bridge.Superblock;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import static org.bitcoinj.core.Utils.HEX;
